refactor: enhance CollectionListItem options
